xpfe/appshell/test/test_windowlessBrowser.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=815847
-->
<window title="Mozilla Bug 815847"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

<body  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1214174">Mozilla Bug 1214174</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>

<script class="testbody" type="application/javascript">
<![CDATA[

var Cu = Components.utils;
var Ci = Components.interfaces;

Cu.import('resource://gre/modules/Services.jsm');

function testWindowlessBrowser(chromePrivileged) {
  var webNav = Services.appShell.createWindowlessBrowser(chromePrivileged);

  ok(webNav, "createWindowlessBrowser should return a wevNav");

  let interfaceRequestor = webNav.QueryInterface(Ci.nsIInterfaceRequestor);
  let docShell = interfaceRequestor.getInterface(Ci.nsIDocShell);

  ok(docShell, "docShell should be defined");
  ok(docShell.contentViewer.DOMDocument.defaultView, "docShell defaultView should be defined");

  var win = docShell.contentViewer.DOMDocument.defaultView;

  ok(win.screenX == 0, "window.screenX should be 0 in a windowless browser");
  ok(win.screenY == 0, "window.screenY should be 0 in a windowless browser");
  ok(win.outerWidth == 0, "window.outerWidth should be 0 in a windowless browser");
  ok(win.outerHeight == 0, "window.outerHeight should be 0 in a windowless browser");
  
  ok(win.sidebar, "window.sidebar should be defined");
  ok(win.external, "window.external should be defined");
  
  var exception;

  try {
    win.external.AddSearchProvider("http://test-fake.url");
  } catch(e) {
    exception = e;
  }
  
  ok(!exception, "window.external.AddSearchProvider should be ignore withour raising an exception");
}

info("Test Bug 1214174 on a content privileged windowless browser");
testWindowlessBrowser(false);

info("Test Bug 1214174 on a chrome privileged windowless browser");
testWindowlessBrowser(true);

]]>
</script>

</window>
